stage 4

+++ src/main/java/Main.java
@@ -40,7 +40,7 @@ public class Main {
       return bencodedString.substring(firstColonIndex+1, firstColonIndex+1+length);
     } else if (bencodedString.startsWith("i") && bencodedString.endsWith("e")) {
       String integerString = bencodedString.substring(1, bencodedString.length() - 1);
-      return String.valueOf(Integer.parseInt(integerString));
+      return String.valueOf(Long.parseLong(integerString));
     } else {
       throw new RuntimeException("Only strings and integers are supported at the moment");
     }
